+++ plone/app/layout/analytics/view.py
@@ -6,10 +6,13 @@ from zope.interface import implements
 from zope.viewlet.interfaces import IViewlet
 
 from Products.Five.browser import BrowserView
+from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 
 
 class AnalyticsViewlet(BrowserView):
     implements(IViewlet)
+
+    render = ViewPageTemplateFile("view.pt")
 
     def __init__(self, context, request, view, manager):
         super(AnalyticsViewlet, self).__init__(context, request)
@@ -18,17 +21,14 @@ class AnalyticsViewlet(BrowserView):
         self.request = request
         self.view = view
         self.manager = manager
+        self.webstats_js = ''
 
     def update(self):
-        pass
-
-    def render(self):
         """render the webstats snippet"""
         registry = getUtility(IRegistry)
         site_settings = registry.forInterface(ISiteSchema, prefix="plone", check=False)
         try:
             if site_settings.webstats_js:
-                return site_settings.webstats_js
-            return ''
+                self.webstats_js = site_settings.webstats_js
         except AttributeError:
-            return ''
+            pass